simulator/requirements.txt
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pyyaml>=6.0.1
gunicorn
requests>=2.31.0